requirements.txt
colorama==0.4.6
numpy==1.26.1
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
